nested api. classrooms from instructors

+++ Backend/db/migrate/20190928161749_create_tasks.rb
@@ -3,7 +3,7 @@ class CreateTasks < ActiveRecord::Migration[5.2]
     create_table :tasks do |t|
       t.belongs_to :session
       t.string :name
-      t.decimal :difficulty
+      t.integer :difficulty
     end
   end
 end
