+++ Backend/db/migrate/20190928162145_create_session_results.rb
@@ -4,7 +4,7 @@ class CreateSessionResults < ActiveRecord::Migration[5.2]
       t.belongs_to :session
       t.belongs_to :student
       t.boolean :completed
-      t.decimal :totalScore
+      t.integer :totalScore
     end
   end
 end
